+++ sphinx/source/changelog.rst
@@ -7,6 +7,8 @@ Changelog
   parsing the skool file in :ref:`rsubMode` and :ref:`rfixMode`)
 * Added the ``--data`` option to :ref:`skool2bin.py` (for processing
   :ref:`defb`, :ref:`defs` and :ref:`defw` directives)
+* Added the ``--verbose`` option to :ref:`skool2bin.py` (for showing
+  information on each converted instruction)
 * The ``address`` parameter of the :ref:`defb`, :ref:`defs` and :ref:`defw`
   directives is now optional
 * :ref:`defb`, :ref:`defs` and :ref:`defw` directives in non-entry blocks are